Cpg134 product qa issue fix raj

See merge request IvyProjects/IvyCpg_Product_MobileApp_Android!649

+++ ivycpg/src/main/java/com/ivy/cpg/view/order/OrderHelper.java
@@ -1109,8 +1109,8 @@ public class OrderHelper {
 
         int pieceCount;
         float srp;
-        double csrp;
-        double osrp;
+        float csrp;
+        float osrp;
         int orderPieceQty;
         int orderCaseQty;
         int orderOuterQty;
@@ -2119,8 +2119,8 @@ public class OrderHelper {
 
         int totalqty;
         float srp;
-        double csrp;
-        double osrp;
+        float csrp;
+        float osrp;
         double prodDisc;
         double schemeDisc;
         double taxAmount;
